refactor(example-test-token-v1-registry): move emptychoicecontext
