+++ examples/test-token-v1-registry/src/api/transfer-instruction/common.ts
@@ -2,7 +2,6 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import {
-    ChoiceContext,
     OperationHandler,
     Operations,
 } from '../../openapi-ts/transfer-instruction-v1'
@@ -11,8 +10,3 @@ import { APIOperationHandler } from '../../types'
 export type TransferInstructionAPIHandler<
     operationId extends keyof Operations,
 > = APIOperationHandler<OperationHandler<operationId>>
-
-export const emptyChoiceContext: ChoiceContext = {
-    choiceContextData: {},
-    disclosedContracts: [],
-}
