+++ examples/test-token-v1-registry/src/api/transfer-instruction/getTransferInstructionAcceptContext.ts
@@ -1,7 +1,8 @@
 // Copyright (c) 2025-2026 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-import { emptyChoiceContext, TransferInstructionAPIHandler } from './common'
+import { emptyChoiceContext } from '../common'
+import { TransferInstructionAPIHandler } from './common'
 
 export const getTransferInstructionAcceptContext: TransferInstructionAPIHandler<
     'getTransferInstructionAcceptContext'
